api/requirements-vercel.txt
# Vercel-specific requirements
# This file contains dependencies optimized for Vercel serverless deployment

fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
hdbcli==2.23.24
langchain-core==0.3.15
langchain-hana==0.2.1
python-multipart==0.0.6
rdflib==7.0.0
langchain==0.3.25
numpy>=1.26.0
sentence-transformers>=2.2.2

# Vercel-specific packages
vercel-fastapi-utils>=0.1.0